Support for state machine documentation during initialization
# What

Added optional `DescribeState` and `DescribeTransitionType` methods to
the state machine that users can call to document the state machine
states and transition types during state machine initialization.

Also added `TransitionRuleDoc` field to `TransitionRule` so users can
optionally fill it out with documentation for the transition rule
when they create it.

# Why

When the above methods and fields are used, users can then call `AsJSON`
on the state machine instance to get a JSON that describes all the
states, transition types and transition rules of the state machine. This
JSON can be used by users to generate documentation / graphs for the
state machine

This is probably not too useful for small state machines, but for large
state machines with many states, transition types and transition rules,
it can be very hard for users to follow the state machine logic, and so
encouraging them to document the state machine can be very useful for
maintenace purposes

Every state has a clear purpose, every transition type has a clear
explanation of in which circumstances it runs, each transition rule
has a clear explanation of why it was added and what it's trying to
solve.

# Example

You can see an example of such documentation / JSON here:

https://github.com/omertuc/assisted-service/tree/statemachine/cmd/graphstatemachine